# Mark arrays as collectable at runtime