Arrays in this PR are marked with GC_NOT_COLLECTABLE by default. In arrays, only objects and references can actually contribute to a cycle because cycles with just arrays are impossible to create from userland. Once an object or reference is added to the array the GC_NOT_COLLECTABLE flag is removed. This patch does not re-add the flag when the potentially cyclic element is removed. This would be possible but would require tracking the number of elements on override/remove.

Benchmarks are yet to be done. This touches several hot-paths so we'll need to make sure it's worth it. JIT also needs an adjustment in one of the handlers which I'll look into. @dstogov I'd be very glad to hear your general thoughts before continuing to avoid working on something potentially non-feasible.

---

Btw, something similar might be possible for objects once [dynamic properties](https://wiki.php.net/rfc/deprecate_dynamic_properties) gets fully removed.

When:

* all properties of a class are typed
* and all of those types are non cyclic, meaning anything but:
    * arrays
    * references to other cyclic classes

then the object of the given class doesn't need to be added to the gc buffer. The two could also be combined, meaning non-cyclic classes added to arrays can't make the arrays cyclic.