@iluuu1994 2022-11-24
I tried benchmarking with Valgrind as well as with `perf stat` but the results seem unreliable. The results vary wildly between runs and thus it's hard to tell if the change has a positive or negative effect, if any at all. Tips on how to get more accurate measurements would be appreciated.

<details>

**Symfony Demo /**

> ZEND_DONT_UNLOAD_MODULES=1 valgrind --tool=callgrind --separate-recs=1 --dump-instr=yes --cache-sim=no ../php-src/sapi/cgi/php-cgi -d zend_extension=$(pwd)/../php-src/modules/opcache.so -d opcache.enable=1 -T100 public/index.php > /dev/null

```
Before:
Elapsed time: 14.515504 sec
==504704== 
==504704== Events    : Ir
==504704== Collected : 3090273121
==504704== 
==504704== I   refs:      3,090,273,121

After:
Elapsed time: 14.826476 sec
==571964== 
==571964== Events    : Ir
==571964== Collected : 3099881221
==571964== 
==571964== I   refs:      3,099,881,221
```

**Symfony Demo /de/blog/**:

> ZEND_DONT_UNLOAD_MODULES=1 valgrind --tool=callgrind --separate-recs=1 --dump-instr=yes --cache-sim=no ../php-src/sapi/cgi/php-cgi -d zend_extension=$(pwd)/../php-src/modules/opcache.so -d opcache.enable=1 -T100 public/index.php > /dev/null

```
Before:
Elapsed time: 45.554749 sec
==504502== 
==504502== Events    : Ir
==504502== Collected : 10604005060
==504502== 
==504502== I   refs:      10,604,005,060

After:
Elapsed time: 45.193365 sec
==440372== 
==440372== Events    : Ir
==440372== Collected : 10597199515
==440372== 
==440372== I   refs:      10,597,199,515
```

**PHPStan**

```
Before:
 Performance counter stats for '../php-src/sapi/cli/php -d memory_limit=2G -d zend_extension=/home/ilutov/Developer/phpstan-src/../php-src/modules/opcache.so -d opcache.enable_cli=1 bin/phpstan --debug':

        343,692.03 msec task-clock:u                     #    1.000 CPUs utilized          
                 0      context-switches:u               #    0.000 /sec                   
                 0      cpu-migrations:u                 #    0.000 /sec                   
           139,033      page-faults:u                    #  404.528 /sec                   
 1,183,713,485,667      cycles:u                         #    3.444 GHz                      (50.00%)
 1,765,963,438,942      instructions:u                   #    1.49  insn per cycle           (62.50%)
   177,913,785,437      branches:u                       #  517.655 M/sec                    (62.49%)
     3,210,747,790      branch-misses:u                  #    1.80% of all branches          (62.50%)
   649,128,225,069      L1-dcache-loads:u                #    1.889 G/sec                    (62.50%)
    17,876,842,843      L1-dcache-load-misses:u          #    2.75% of all L1-dcache accesses  (62.51%)
     3,982,013,044      LLC-loads:u                      #   11.586 M/sec                    (50.00%)
       360,078,544      LLC-load-misses:u                #    9.04% of all LL-cache accesses  (50.00%)

     343.785708808 seconds time elapsed

     339.094749000 seconds user
       0.827771000 seconds sys

After:
 Performance counter stats for '../php-src/sapi/cli/php -d memory_limit=2G -d zend_extension=/home/ilutov/Developer/phpstan-src/../php-src/modules/opcache.so -d opcache.enable_cli=1 bin/phpstan --debug':

        336,441.28 msec task-clock:u                     #    1.000 CPUs utilized          
                 0      context-switches:u               #    0.000 /sec                   
                 0      cpu-migrations:u                 #    0.000 /sec                   
           140,104      page-faults:u                    #  416.429 /sec                   
 1,162,996,306,873      cycles:u                         #    3.457 GHz                      (50.00%)
 1,768,527,792,843      instructions:u                   #    1.52  insn per cycle           (62.50%)
   178,133,087,555      branches:u                       #  529.463 M/sec                    (62.49%)
     3,187,759,190      branch-misses:u                  #    1.79% of all branches          (62.50%)
   650,433,979,507      L1-dcache-loads:u                #    1.933 G/sec                    (62.50%)
    17,495,596,566      L1-dcache-load-misses:u          #    2.69% of all L1-dcache accesses  (62.51%)
     3,987,081,054      LLC-loads:u                      #   11.851 M/sec                    (50.00%)
       304,078,180      LLC-load-misses:u                #    7.63% of all LL-cache accesses  (50.00%)

     336.497944515 seconds time elapsed

     331.160239000 seconds user
       1.020707000 seconds sys
```
</details>