@dstogov 2022-11-28
It looks like the current implementation doesn't make significant effect. Some apps become a bit faster others are slower.

I think we should postpone this and think about non-collectable objects first. After disabling creation of dynamic object properties (they are deprecated in PHP-8.2)  we may determine non-collectable classes just analysing their properties.

Then it would make sense to adopt ``gc_ht_value_may_cause_cycle()``  to support for non-collectable objects.

Also the GC algorithm might be improved. See the related ideas described at http://www.eng.usf.edu/~chang5/papers/12/RC_chang_12.pdf